client/openapi/py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aju"
version = "0.1.0"
description = "Official Python SDK for the aju HTTP API. Generated from client/openapi/openapi.yaml."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "See repository LICENSE" }
dependencies = [
  "httpx>=0.27,<1.0",
  "attrs>=23.2",
  "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
  "openapi-python-client>=0.21,<1.0",
  "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["aju"]

[tool.ruff]
line-length = 100
target-version = "py310"
